+++ ExamService/pkg/repository/repository.go
@@ -23,15 +23,15 @@ type DatabaseRepository interface {
 	FindExamsByUserIdOrderByUpdateAtDesc(
 		ctx context.Context,
 		userId string,
-		skip, limit int64,
+		skip, limit int32,
 	) (exams []model.Exam, err error)
 	FindExamsByUserIdAndIsPublicOrderByUpdateAtDesc(
 		ctx context.Context,
 		userId string,
 		isPublic bool,
 	) (exams []model.Exam, err error)
-	DeleteExamById(ctx context.Context, examId string) (deletedCount int64, err error)
-	CountExamsByUserId(ctx context.Context, userId string) (count int64, err error)
+	DeleteExamById(ctx context.Context, examId string) (deletedCount int32, err error)
+	CountExamsByUserId(ctx context.Context, userId string) (count int32, err error)
 
 	// Question
 	CreateQuestion(ctx context.Context, question model.Question) (questionId string, err error)
@@ -40,39 +40,39 @@ type DatabaseRepository interface {
 	FindQuestionsByExamIdOrderByUpdateAtDesc(
 		ctx context.Context,
 		examId string,
-		skip, limit int64,
+		skip, limit int32,
 	) (questions []model.Question, err error)
-	DeleteQuestionById(ctx context.Context, questionId string) (deletedCount int64, err error)
-	DeleteQuestionsByExamId(ctx context.Context, examId string) (deletedCount int64, err error)
+	DeleteQuestionById(ctx context.Context, questionId string) (deletedCount int32, err error)
+	DeleteQuestionsByExamId(ctx context.Context, examId string) (deletedCount int32, err error)
 	CountQuestionsByExamId(
 		ctx context.Context,
 		examId string,
-	) (count int64, err error)
+	) (count int32, err error)
 
 	// AnswerWrong
 	DeleteAnswerWrongsByQuestionId(
 		ctx context.Context,
 		questionId string,
-	) (deletedCount int64, err error)
-	DeleteAnswerWrongsByExamId(ctx context.Context, examId string) (deletedCount int64, err error)
+	) (deletedCount int32, err error)
+	DeleteAnswerWrongsByExamId(ctx context.Context, examId string) (deletedCount int32, err error)
 	UpsertAnswerWrongByTimesPlusOne(
 		ctx context.Context,
 		examId, questionId, userId string,
-	) (modifiedCount, upsertedCount int64, err error)
+	) (modifiedCount, upsertedCount int32, err error)
 
 	// ExamRecord
 	CreateExamRecord(
 		ctx context.Context,
 		examRecord model.ExamRecord,
 	) (examRecordId string, err error)
-	DeleteExamRecordsByExamId(ctx context.Context, examId string) (deletedCount int64, err error)
+	DeleteExamRecordsByExamId(ctx context.Context, examId string) (deletedCount int32, err error)
 	FindExamRecordsByExamIdAndUserIdOrderByUpdateAtDesc(
 		ctx context.Context,
 		examId, userId string,
-		skip, limit int64,
+		skip, limit int32,
 	) (examRecords []model.ExamRecord, err error)
 	CountExamRecordsByExamIdAndUserId(
 		ctx context.Context,
 		examId, userId string,
-	) (count int64, err error)
+	) (count int32, err error)
 }
